requirements.txt
rasa==3.6.21
openai==0.28.1
python-dotenv==1.0.1
sqlalchemy<2.0
aiohttp==3.9.5
requests==2.32.3
colorama==0.4.6
pydantic==1.10.9
PyYAML==6.0.2
streamlit==1.38.0